fix dependency cycle caused by utilDisplayLabel importing presetManager
