+++ modules/util/util.js
@@ -1,7 +1,6 @@
 import { remove as removeDiacritics } from 'diacritics';
 import { fixRTLTextForSvg, rtlRegex } from './svg_paths_rtl_fix';
 
-import { presetManager } from '../presets';
 import { t, localizer } from '../core/localizer';
 import { utilArrayUnion } from './array';
 import { utilDetect } from './detect';
@@ -241,32 +240,6 @@ export function utilDisplayType(id) {
         w: t('inspector.way'),
         r: t('inspector.relation')
     }[id.charAt(0)];
-}
-
-
-// `utilDisplayLabel`
-// Returns a string suitable for display
-// By default returns something like name/ref, fallback to preset type, fallback to OSM type
-//   "Main Street" or "Tertiary Road"
-// If `verbose=true`, include both preset name and feature name.
-//   "Tertiary Road Main Street"
-//
-export function utilDisplayLabel(entity, graphOrGeometry, verbose) {
-    var result;
-    var displayName = utilDisplayName(entity);
-    var preset = typeof graphOrGeometry === 'string' ?
-        presetManager.matchTags(entity.tags, graphOrGeometry) :
-        presetManager.match(entity, graphOrGeometry);
-    var presetName = preset && (preset.suggestion ? preset.subtitle() : preset.name());
-
-    if (verbose) {
-        result = [presetName, displayName].filter(Boolean).join(' ');
-    } else {
-        result = displayName || presetName;
-    }
-
-    // Fallback to the OSM type (node/way/relation)
-    return result || utilDisplayType(entity.id);
 }
