+++ modules/validations/mutually_exclusive_tags.js
@@ -1,6 +1,6 @@
 import { actionChangeTags } from '../actions/change_tags';
 import { t } from '../core/localizer';
-import { utilDisplayLabel } from '../util';
+import { utilDisplayLabel } from '../util/utilDisplayLabel';
 import { validationIssue, validationIssueFix } from '../core/validation';
 import { osmMutuallyExclusiveTagPairs } from '../osm/tags';
