+++ modules/validations/private_data.js
@@ -1,6 +1,7 @@
 import { actionChangeTags } from '../actions/change_tags';
 import { t } from '../core/localizer';
-import { utilDisplayLabel, utilTagDiff } from '../util';
+import { utilTagDiff } from '../util';
+import { utilDisplayLabel } from '../util/utilDisplayLabel';
 import { validationIssue, validationIssueFix } from '../core/validation';
